Split up RoutesPanel file
