+++ studio/src/pages/RequestorPage/NavigationPanel/RoutesPanel/index.tsx
@@ -1,0 +1,2 @@
+export { RoutesPanel } from "./RoutesPanel";
+export { RouteItem } from "./RouteItem";
