+++ studio/src/pages/RequestorPage/NavigationPanel/index.ts
@@ -1,2 +1,2 @@
 // In the future we'll have a navigation panel (which will render the RoutesPanel)
-export { RoutesPanel } from "./RoutesPanel";
+export { RoutesPanel, RouteItem } from "./RoutesPanel";
